new file: api functions
